requirements.txt
streamlit
opencv-python
numpy
pyyaml
pillow
matplotlib
scipy
ultralytics
librosa
soundfile
